Attempt to make dtype conversions more centralized
Aims to solve the following:
1. dtype conversions were scattered in multiple places (`datatypes.py`, `series.py`, `frame.py`), now it is all in `datatypes.py`. Note that py_type_to_dtype` still supports a Polars dtype as input, as it is used as such in multiple places. Strictly speaking it should only accept Python types.
2. use dicts rather than long if-else trees
3. break the circular import by factoring out the datatypes for `construction.py` into `datatypes_constructor.py`. Should this be moved into `construction.py`? I think this also removes the need for the `if DOCUMENTING` clause, but I may be wrong here.
